Several changes requested in the PR comments
https://github.com/OpenCircuits/OpenCircuits/pull/196
Replaced getters/setters with methods
MainDesignerController.GetCamera instead of CanvasToWorld and CanvasToScreen
Selection popup modules each in their own file
Import statements do not have spaces padding the brackets
