+++ site/public/ts/Main.ts
@@ -7,6 +7,7 @@ import {HeaderController} from "./controllers/HeaderController";
 import {ItemNavController} from "./controllers/ItemNavController";
 // import {InputController} from "./utils/input/InputController";
 import {ContextMenuController} from "./controllers/ContextMenuController";
+import {SelectionPopupController} from "./controllers/SelectionPopupController";
 
 function Start() {
     Load(Init);
@@ -22,6 +23,7 @@ function Init() {
     HeaderController.Init(MainDesignerController.GetDesigner());
     ItemNavController.Init(MainDesignerController.GetDesigner());
     ContextMenuController.Init();
+    SelectionPopupController.Init(MainDesignerController.GetCamera());
     // ICDesignerController.Init();
 
     MainDesignerController.Render();
