+++ site/public/ts/controllers/SelectionPopupController.ts
@@ -1,36 +1,43 @@
-import { Vector } from "../utils/math/Vector";
-import * as Modules from "../utils/selectionpopup/SelectionPopupModules";
-import { MainDesignerController } from "./MainDesignerController";
-import { Component } from "../models/ioobjects/Component";
+import {Vector} from "../utils/math/Vector";
+import {MainDesignerController} from "./MainDesignerController";
+import {Component} from "../models/ioobjects/Component";
+import {Camera} from "../utils/Camera";
+import {SelectionPopupModule} from "../utils/selectionpopup/SelectionPopupModule";
+import {TitlePopupModule} from "../utils/selectionpopup/TitlePopupModule";
+import {PositionPopupModule} from "../utils/selectionpopup/PositionPopupModule";
 
 /**
  * A popup that exposes certain properties of the selected components to the user
- * ! Controls its own DOM element
+ * ! Controls its own DOM element(s)
+ * TODO: use decorators or some other interface to determine what properties are available
  */
 class SelectionPopupController {
+    private camera: Camera;
     private div: HTMLDivElement;
-    private modules: Array<Modules.SelectionPopupModule>;
+    private modules: Array<SelectionPopupModule>;
     private pos: Vector;
 
-    constructor(div_id: string = "popup") {
+    constructor(camera: Camera, div_id: string = "popup") {
+        this.camera = camera;
+
         this.div = document.getElementById(div_id) as HTMLDivElement;
         // ? .js sets position to "absolute" -- why? Why not set in the css file
 
-        this.modules = new Array<Modules.SelectionPopupModule>(
-            new Modules.TitlePopupModule(this.div),
-            new Modules.PositionPopupModule(this.div),
+        this.modules = new Array<SelectionPopupModule>(
+            new TitlePopupModule(this.div),
+            new PositionPopupModule(this.div),
         );
         this.pos = new Vector(0, 0);
     }
 
-    set position(v: Vector) {
+    setPos(v: Vector) {
         this.pos = v;
 
         this.div.style.left = `${this.pos.x}px`;
         this.div.style.top = `${this.pos.y}px`;
     }
 
-    get position(): Vector {
+    getPos(): Vector {
         return this.pos;
     }
 
@@ -52,14 +59,13 @@ class SelectionPopupController {
                     count += 1;
                 }
             }
-            let screen_pos = MainDesignerController.CanvasToScreen(sum.scale(1/count));
+            let screen_pos = this.camera.getScreenPos(sum.scale(1/count));
             //console.log(this.div.clientHeight, document.body.clientHeight);
             screen_pos.y = screen_pos.y - (this.div.clientHeight/2);
             // TODO: clamp should make sure not to overlap with other screen elements
             //const lo = new Vector(0);
             //const hi = new Vector(document.body.clientWidth, document.body.clientHeight);
-            this.position = screen_pos;// Vector.clamp(screen_pos, lo, hi);
-
+            this.setPos(screen_pos);// Vector.clamp(screen_pos, lo, hi);
 
             this.show();
         } else {
@@ -78,16 +84,22 @@ class SelectionPopupController {
 }
 
 namespace Singleton {
-    const p = new SelectionPopupController();
+    // readonly doesn't work in namespaces :(
+    let popup: SelectionPopupController;
 
+
+    export function Init(camera: Camera) {
+        console.assert(popup === undefined);
+        popup = new SelectionPopupController(camera);
+    }
     export function Update() {
-        p.update();
+        popup.update();
     }
     export function Show() {
-        p.show();
+        popup.show();
     }
     export function Hide() {
-        p.hide();
+        popup.hide();
     }
 }
